c-decl.c (link_hash_hash): Avoid warning about casting pointer to integer of different size.
	* c-decl.c (link_hash_hash): Avoid warning about casting pointer
	to integer of different size.

From-SVN: r69302

+++ gcc/c-decl.c
@@ -6794,7 +6794,7 @@ static hashval_t
 link_hash_hash (const void *x_p)
 {
   tree x = (tree)x_p;
-  return (hashval_t) DECL_ASSEMBLER_NAME (x);
+  return (hashval_t) (long)DECL_ASSEMBLER_NAME (x);
 }
 
 static int
